backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
sentence-transformers==2.2.2
openai==1.3.7
python-dotenv==1.0.0